update compass-crud to shared config, still has a few eslint issues
