--- packages/compass-crud/.browserslistrc
@@ -1,1 +1,0 @@
-Electron >= 6
