--- packages/compass-crud/.npmignore
@@ -1,19 +1,0 @@
-.babelrc
-.eslintrc
-.editorconfig
-.nycrc
-.travis.yml
-config/
-electron/
-src/
-mocha-webpack.opts
-scripts/
-test/
-coverage/
-.nyc_output/
-.github/
-lib/index.html
-karma.conf.js
-
-scripts/**/*
-!scripts/download-akzidenz.js
