+++ packages/compass-crud/package.json
@@ -3,39 +3,62 @@
   "productName": "Compass CRUD plugin",
   "version": "12.24.0",
   "description": "Compass Plugin for CRUD Operations",
+  "author": {
+    "name": "MongoDB Inc",
+    "email": "compass@mongodb.com"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/mongodb-js/compass.git"
   },
-  "main": "lib/index.js",
+  "license": "SSPL",
+  "homepage": "https://github.com/mongodb-js/compass",
+  "bugs": {
+    "url": "https://jira.mongodb.org/projects/COMPASS/issues",
+    "email": "compass@mongodb.com"
+  },
+  "files": [
+    "dist"
+  ],
+  "main": "dist/index.js",
   "compass:main": "src/index.js",
+  "exports": {
+    "require": "./dist/index.js",
+    "browser": "./dist/browser.js"
+  },
   "compass:exports": {
     ".": "./src/index.js"
   },
-  "apiVersion": "3.0.0",
   "scripts": {
-    "clean": "rimraf lib",
-    "precompile": "npm run clean",
-    "compile": "cross-env NODE_ENV=production webpack --config ./config/webpack.prod.config.js",
-    "start": "cross-env NODE_ENV=development webpack-dev-server --config ./config/webpack.dev.config.js",
-    "start:watch": "npm run clean && webpack --config ./config/webpack.watch.config.js",
-    "pretest": "mongodb-runner start --port 27018",
-    "test": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\"",
-    "test:watch": "cross-env NODE_ENV=test mocha-webpack \"./src/**/*.spec.js\" --watch",
-    "posttest": "mongodb-runner stop --port 27018",
-    "cover": "nyc npm run test",
-    "test-check-ci": "npm run check && npm run cover",
-    "depcheck": "depcheck",
-    "lint": "eslint \"./src/**/*.{js,jsx}\" \"./test/**/*.js\" \"./electron/**/*.js\" \"./config/**/*.{js,jsx}\"",
-    "check": "npm run depcheck && npm run lint",
     "prepublishOnly": "npm run compile && compass-scripts check-exports-exist",
-    "test-ci": "npm run test",
+    "compile": "npm run webpack -- --mode production",
+    "prewebpack": "rimraf ./dist",
+    "webpack": "webpack-compass",
+    "start": "npm run webpack serve -- --mode development",
+    "analyze": "npm run webpack -- --mode production --analyze",
+    "typecheck": "tsc --noEmit",
+    "eslint": "eslint",
+    "prettier": "prettier",
+    "lint": "npm run eslint .",
+    "prettier-skipped-for-now": "&& npm run prettier -- --check .",
+    "depcheck": "depcheck",
+    "check": "npm run typecheck && npm run lint && npm run depcheck",
+    "check-ci": "npm run check",
+    "pretest": "mongodb-runner start --port 27018",
+    "test": "mocha",
+    "posttest": "mongodb-runner stop --port 27018",
+    "test-electron": "xvfb-maybe electron-mocha --no-sandbox",
+    "test-cov": "nyc -x \"**/*.spec.*\" --reporter=lcov --reporter=text --reporter=html npm run test",
+    "test-watch": "npm run test -- --watch",
+    "test-ci": "npm run test-cov",
     "posttest-ci": "node ../../scripts/killall-mongo.js",
-    "bootstrap": "npm run compile"
+    "test-ci-electron": "npm run test-electron",
+    "reformat": "npm run prettier -- --write ."
   },
-  "license": "SSPL",
   "peerDependencies": {
-    "@hot-loader/react-dom": "^16.9.0",
     "@mongodb-js/compass-components": "^0.15.0",
     "@mongodb-js/compass-query-bar": "^7.22.0",
     "bson": "*",
@@ -47,97 +70,50 @@
     "react-tooltip": "^3.11.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.3",
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-decorators": "^7.14.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/register": "^7.13.16",
-    "@hot-loader/react-dom": "^16.9.0",
     "@mongodb-js/compass-components": "^0.15.0",
     "@mongodb-js/compass-query-bar": "^7.22.0",
-    "autoprefixer": "^9.4.6",
-    "babel-loader": "^8.2.2",
+    "@mongodb-js/eslint-config-compass": "^0.9.0",
+    "@mongodb-js/mocha-config-compass": "^0.10.0",
+    "@mongodb-js/prettier-config-compass": "^0.5.0",
+    "@mongodb-js/tsconfig-compass": "^0.6.0",
+    "@mongodb-js/webpack-config-compass": "^0.9.0",
     "chai": "^4.1.2",
     "chai-as-promised": "*",
-    "chai-enzyme": "1.0.0-beta.1",
     "classnames": "^2.2.6",
-    "core-js": "^3.12.1",
-    "cross-env": "^7.0.0",
-    "css-loader": "^4.3.0",
     "depcheck": "^1.4.1",
     "electron": "^13.5.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^7.25.0",
-    "eslint-config-mongodb-js": "^5.0.3",
-    "eslint-plugin-react": "^7.24.0",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "file-loader": "^5.1.0",
-    "font-awesome": "^4.7.0",
     "hadron-app": "^4.23.0",
     "hadron-app-registry": "^8.12.0",
     "hadron-react-buttons": "^5.9.0",
     "hadron-react-components": "^5.15.0",
-    "html-webpack-plugin": "^3.2.0",
-    "ignore-loader": "^0.1.2",
-    "istanbul-instrumenter-loader": "^3.0.1",
-    "jquery": "^3.0.0",
-    "jsdom": "^16.7.0",
-    "jsdom-global": "^3.0.2",
-    "less": "^3.11.1",
-    "less-loader": "^7.3.0",
     "mocha": "^5.0.0",
-    "mocha-webpack": "^2.0.0-beta.0",
     "mongodb-connection-model": "^21.17.0",
     "mongodb-data-service": "^21.21.0",
     "mongodb-ns": "^2.4.0",
     "mongodb-runner": "^4.9.0",
-    "node-loader": "^0.6.0",
-    "numeral": "^2.0.6",
     "nyc": "^15.0.0",
-    "peer-deps-externals-webpack-plugin": "^1.0.4",
-    "pluralize": "^8.0.0",
-    "postcss-loader": "^2.1.6",
     "prop-types": "^15.7.2",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-fontawesome": "^1.6.1",
-    "react-hot-loader": "^4.13.0",
     "reflux": "^0.4.1",
     "reflux-state-mixin": "github:mongodb-js/reflux-state-mixin",
     "rimraf": "^3.0.0",
-    "semver": "^6.3.0",
-    "shebang-loader": "^0.0.1",
     "sinon": "^8.1.1",
-    "sinon-chai": "^3.4.0",
-    "style-loader": "^2.0.0",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.46.0",
-    "webpack-bundle-analyzer": "^3.1.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^4.2.2",
-    "webpack-node-externals": "^3.0.0"
+    "webpack": "^4.46.0"
   },
   "dependencies": {
     "@mongodb-js/compass-logging": "^0.12.0",
     "ag-grid-community": "20.2.0",
     "ag-grid-react": "20.2.0",
     "bson": "*",
-    "debug": "4.3.0",
     "fast-json-parse": "^1.0.3",
     "hadron-document": "^7.11.0",
     "hadron-type-checker": "^6.9.0",
     "lodash": "^4.17.15",
-    "moment": "^2.27.0",
     "react-click-outside": "^3.0.1",
     "react-tooltip": "^3.11.1"
-  },
-  "homepage": "https://github.com/mongodb-js/compass",
-  "bugs": {
-    "url": "https://jira.mongodb.org/projects/COMPASS/issues",
-    "email": "compass@mongodb.com"
   }
 }
