+++ packages/compass-crud/src/components/table-view/types.jsx
@@ -140,7 +140,9 @@ class Types extends React.Component {
         <li key={type}>
           <span
             className={`editable-element-type-${type.toLowerCase()}`}
-            onMouseDown={this.handleTypeChange.bind(this)}>
+            onMouseDown={this.handleTypeChange.bind(this)}
+            role="button"
+          >
             {type}
           </span>
         </li>
