+++ packages/compass-crud/src/stores/grid-store.spec.js
@@ -1,9 +1,9 @@
-/* eslint guard-for-in: 0 no-loop-func: 0 */
+import { assert, expect } from 'chai';
+
 import configureStore from '../../src/stores/grid-store';
 import configureActions from '../../src/actions';
-import { assert } from 'chai';
 
-describe('store', () => {
+describe('store', function() {
   let store;
   const actions = configureActions();
   const showing = {
@@ -26,20 +26,20 @@ describe('store', () => {
     0: 'Mixed', 1: 'Mixed', 2: 'Int32'
   };
 
-  describe('#columns', () => {
-    describe('adding an array column', () => {
-      describe('editOnly=true', () => {
-        before((done) => {
+  describe('#columns', function() {
+    describe('adding an array column', function() {
+      describe('editOnly=true', function() {
+        before(function(done) {
           store = configureStore({ actions: actions });
           store.resetColumns(arrayCols);
           expect(store.columns).to.deep.equal(arrayCols);
           expect(store.showing).to.deep.equal(arrayShowing);
           done();
         });
-        after((done) => {
+        after(function(done) {
           done();
         });
-        it('does not trigger with add', (done) => {
+        it('does not trigger with add', function(done) {
           const unsubscribe = store.listen((params) => {
             unsubscribe();
             expect(params).to.deep.equal({
@@ -50,31 +50,31 @@ describe('store', () => {
           });
           store.addColumn(1, 0, 3, ['array'], true, true, 'id2');
         });
-        it('updates this.columns correctly', () => {
+        it('updates this.columns correctly', function() {
           expect(store.columns).to.deep.equal({
             0: {id1: 'String', id2: 'Int64'},
             1: {id1: 'Int64'},
             2: {id1: 'Int32', id2: 'Double'}
           });
         });
-        it('updates this.showing correctly', () => {
+        it('updates this.showing correctly', function() {
           expect(store.showing).to.deep.equal({
             0: 'Mixed', 1: 'Int64', 2: 'Mixed'
           });
         });
       });
-      describe('editOnly=false', () => {
-        before((done) => {
+      describe('editOnly=false', function() {
+        before(function(done) {
           store = configureStore({ actions: actions });
           store.resetColumns(arrayCols);
           expect(store.columns).to.deep.equal(arrayCols);
           expect(store.showing).to.deep.equal(arrayShowing);
           done();
         });
-        after((done) => {
+        after(function(done) {
           done();
         });
-        it('does trigger with add', (done) => {
+        it('does trigger with add', function(done) {
           const unsubscribe = store.listen((params) => {
             unsubscribe();
             expect(params).to.deep.equal({
@@ -92,7 +92,7 @@ describe('store', () => {
           });
           store.addColumn(1, 0, 3, ['array'], true, false, 'id1');
         });
-        it('updates this.columns correctly', () => {
+        it('updates this.columns correctly', function() {
           expect(store.columns).to.deep.equal({
             0: {id1: 'String', id2: 'Int64'},
             1: {id2: 'Double'},
@@ -100,16 +100,16 @@ describe('store', () => {
             3: {id1: 'Int32'}
           });
         });
-        it('updates this.showing correctly', () => {
+        it('updates this.showing correctly', function() {
           expect(store.showing).to.deep.equal({
             0: 'Mixed', 1: 'Double', 2: 'Int64', 3: 'Int32'
           });
         });
       });
-      describe('with elements marked as removed', () => {
-        describe('editOnly=false', () => {
-          describe('mark removed is same element as added', () => {
-            before((done) => {
+      describe('with elements marked as removed', function() {
+        describe('editOnly=false', function() {
+          describe('mark removed is same element as added', function() {
+            before(function(done) {
               store = configureStore({ actions: actions });
               store.resetColumns(arrayCols);
               store.elementMarkRemoved(1, 'id1');
@@ -126,10 +126,10 @@ describe('store', () => {
               });
               done();
             });
-            after((done) => {
+            after(function(done) {
               done();
             });
-            it('does trigger with add', (done) => {
+            it('does trigger with add', function(done) {
               const unsubscribe = store.listen((params) => {
                 unsubscribe();
                 expect(params).to.deep.equal({
@@ -147,26 +147,26 @@ describe('store', () => {
               });
               store.addColumn(1, 0, 3, ['array'], true, false, 'id1');
             });
-            it('updates this.columns correctly', () => {
+            it('updates this.columns correctly', function() {
               expect(store.columns).to.deep.equal({
                 0: {id1: 'String', id2: 'Int64'},
                 1: {id2: 'Double'},
                 3: {id1: 'Int32'}
               });
             });
-            it('updates this.showing correctly', () => {
+            it('updates this.showing correctly', function() {
               expect(store.showing).to.deep.equal({
                 0: 'Mixed', 1: 'Double', 2: 'Int32', 3: 'Int32'
               });
             });
-            it('updates this.stageRemoved correctly', () => {
+            it('updates this.stageRemoved correctly', function() {
               expect(store.stageRemove).to.deep.equal({
                 2: {id1: true}
               });
             });
           });
-          describe('mark removed is before added', () => {
-            before((done) => {
+          describe('mark removed is before added', function() {
+            before(function(done) {
               store = configureStore({ actions: actions });
               store.resetColumns(arrayCols);
               store.elementMarkRemoved(0, 'id1');
@@ -183,10 +183,10 @@ describe('store', () => {
               });
               done();
             });
-            after((done) => {
+            after(function(done) {
               done();
             });
-            it('does trigger with add', (done) => {
+            it('does trigger with add', function(done) {
               const unsubscribe = store.listen((params) => {
                 unsubscribe();
                 expect(params).to.deep.equal({
@@ -204,7 +204,7 @@ describe('store', () => {
               });
               store.addColumn(1, 0, 3, ['array'], true, false, 'id1');
             });
-            it('updates this.columns correctly', () => {
+            it('updates this.columns correctly', function() {
               expect(store.columns).to.deep.equal({
                 0: {id2: 'Int64'},
                 1: {id2: 'Double'},
@@ -212,19 +212,19 @@ describe('store', () => {
                 3: {id1: 'Int32'}
               });
             });
-            it('updates this.showing correctly', () => {
+            it('updates this.showing correctly', function() {
               expect(store.showing).to.deep.equal({
                 0: 'Int64', 1: 'Double', 2: 'Int64', 3: 'Int32'
               });
             });
-            it('updates this.stageRemoved correctly', () => {
+            it('updates this.stageRemoved correctly', function() {
               expect(store.stageRemove).to.deep.equal({
                 0: {id1: true}
               });
             });
           });
-          describe('mark removed is after added', () => {
-            before((done) => {
+          describe('mark removed is after added', function() {
+            before(function(done) {
               store = configureStore({ actions: actions });
               store.resetColumns(arrayCols);
               store.elementMarkRemoved(2, 'id1');
@@ -240,10 +240,10 @@ describe('store', () => {
               });
               done();
             });
-            after((done) => {
+            after(function(done) {
               done();
             });
-            it('does trigger with add', (done) => {
+            it('does trigger with add', function(done) {
               const unsubscribe = store.listen((params) => {
                 unsubscribe();
                 expect(params).to.deep.equal({
@@ -261,19 +261,19 @@ describe('store', () => {
               });
               store.addColumn(1, 0, 3, ['array'], true, false, 'id1');
             });
-            it('updates this.columns correctly', () => {
+            it('updates this.columns correctly', function() {
               expect(store.columns).to.deep.equal({
                 0: {id1: 'String', id2: 'Int64'},
                 1: {id2: 'Double'},
                 2: {id1: 'Int64'}
               });
             });
-            it('updates this.showing correctly', () => {
+            it('updates this.showing correctly', function() {
               expect(store.showing).to.deep.equal({
                 0: 'Mixed', 1: 'Double', 2: 'Int64', 3: 'Int32'
               });
             });
-            it('updates this.stageRemoved correctly', () => {
+            it('updates this.stageRemoved correctly', function() {
               expect(store.stageRemove).to.deep.equal({
                 3: {id1: true}
               });
@@ -282,15 +282,15 @@ describe('store', () => {
         });
       });
     });
-    describe('adding a non-array column', () => {
-      before((done) => {
+    describe('adding a non-array column', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         done();
       });
-      after((done) => {
+      after(function(done) {
         done();
       });
-      it('triggers with correct params', (done) => {
+      it('triggers with correct params', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({
@@ -306,8 +306,8 @@ describe('store', () => {
       });
     });
 
-    describe('removing a column', () => {
-      it('triggers with correct params', (done) => {
+    describe('removing a column', function() {
+      it('triggers with correct params', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({remove: {colIds: ['field3']}});
@@ -318,23 +318,23 @@ describe('store', () => {
     });
   });
 
-  describe('#resetCols', () => {
-    before((done) => {
+  describe('#resetCols', function() {
+    before(function(done) {
       store = configureStore({ actions: actions });
       showing.field5 = 'Mixed';
       done();
     });
-    after((done) => {
+    after(function(done) {
       showing.field5 = 'Int32';
       done();
     });
-    it('resets columns to empty', () => {
+    it('resets columns to empty', function() {
       store.resetColumns({});
       expect(store.columns).to.deep.equal({});
       expect(store.showing).to.deep.equal({});
       expect(store.stageRemove).to.deep.equal({});
     });
-    it('resets columns to values', () => {
+    it('resets columns to values', function() {
       store.resetColumns(columns);
       expect(store.columns).to.deep.equal(columns);
       expect(store.showing).to.deep.equal(showing);
@@ -342,14 +342,14 @@ describe('store', () => {
     });
   });
 
-  describe('#elementAdded', () => {
-    before((done) => {
+  describe('#elementAdded', function() {
+    before(function(done) {
       store = configureStore({ actions: actions });
       done();
     });
-    describe('Adding a new column', () => {
+    describe('Adding a new column', function() {
       for (const key in columns) {
-        it('triggers correctly', (done) => {
+        it('triggers correctly', function(done) {
           const unsubscribe = store.listen((params) => {
             unsubscribe();
             const show = {};
@@ -360,7 +360,7 @@ describe('store', () => {
           store.elementAdded(key, columns[key]['1'], '1');
         });
       }
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field1: {1: 'Object'},
           field2: {1: 'Array'},
@@ -369,14 +369,14 @@ describe('store', () => {
           field5: {1: 'Int32'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal(showing);
       });
     });
 
-    describe('Adding to a column that already exists with the same type', () => {
+    describe('Adding to a column that already exists with the same type', function() {
       for (const key in columns) {
-        it('does not trigger', () => {
+        it('does not trigger', function() {
           const unsubscribe = store.listen(() => {
             assert.fail();
           });
@@ -387,7 +387,7 @@ describe('store', () => {
           unsubscribe();
         });
       }
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field1: {1: 'Object', 2: 'Object', 3: 'Object'},
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
@@ -396,14 +396,14 @@ describe('store', () => {
           field5: {1: 'Int32', 2: 'Int32'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         showing.field5 = 'Int32';
         expect(store.showing).to.deep.equal(showing);
         showing.field5 = 'Mixed';
       });
     });
-    describe('Adding to a column that already exists with a new type', () => {
-      it('triggers correctly', (done) => {
+    describe('Adding to a column that already exists with a new type', function() {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({updateHeaders: {showing: {field5: 'Mixed'}}});
@@ -411,25 +411,25 @@ describe('store', () => {
         });
         store.elementAdded('field5', 'String', '3');
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal(columns);
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal(showing);
       });
     });
   });
 
-  describe('#elementRemoved', () => {
-    describe('isArray=false', () => {
-      before((done) => {
+  describe('#elementRemoved', function() {
+    describe('isArray=false', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         store.resetColumns(columns);
         done();
       });
-      describe('Removing from a column that already exists with the same type', () => {
+      describe('Removing from a column that already exists with the same type', function() {
         for (const key in columns) {
-          it('does not trigger', () => {
+          it('does not trigger', function() {
             const unsubscribe = store.listen(() => {
               assert.fail();
             });
@@ -440,7 +440,7 @@ describe('store', () => {
             unsubscribe();
           });
         }
-        it('sets this.columns correctly', () => {
+        it('sets this.columns correctly', function() {
           expect(store.columns).to.deep.equal({
             field1: {3: 'Object'},
             field2: {3: 'Array'},
@@ -449,12 +449,12 @@ describe('store', () => {
             field5: {2: 'Int32', 3: 'String'}
           });
         });
-        it('sets this.showing correctly', () => {
+        it('sets this.showing correctly', function() {
           expect(store.showing).to.deep.equal(showing);
         });
       });
-      describe('Removing from a Mixed column', () => {
-        it('triggers correctly', (done) => {
+      describe('Removing from a Mixed column', function() {
+        it('triggers correctly', function(done) {
           const unsubscribe = store.listen((params) => {
             unsubscribe();
             expect(params).to.deep.equal({updateHeaders: {showing: {field5: 'Int32'}}});
@@ -462,7 +462,7 @@ describe('store', () => {
           });
           store.elementRemoved('field5', '3', false);
         });
-        it('sets this.columns correctly', () => {
+        it('sets this.columns correctly', function() {
           expect(store.columns).to.deep.equal({
             field1: {3: 'Object'},
             field2: {3: 'Array'},
@@ -471,13 +471,13 @@ describe('store', () => {
             field5: {2: 'Int32'}
           });
         });
-        it('sets this.showing correctly', () => {
+        it('sets this.showing correctly', function() {
           showing.field5 = 'Int32';
           expect(store.showing).to.deep.equal(showing);
         });
       });
-      describe('Removing the last item from a column', () => {
-        it('triggers correctly', (done) => {
+      describe('Removing the last item from a column', function() {
+        it('triggers correctly', function(done) {
           const unsubscribe = store.listen((params) => {
             unsubscribe();
             expect(params).to.deep.equal({remove: {colIds: ['field5']}});
@@ -485,7 +485,7 @@ describe('store', () => {
           });
           store.elementRemoved('field5', '2', false);
         });
-        it('sets this.columns correctly', () => {
+        it('sets this.columns correctly', function() {
           expect(store.columns).to.deep.equal({
             field1: {3: 'Object'},
             field2: {3: 'Array'},
@@ -493,26 +493,26 @@ describe('store', () => {
             field4: {3: 'Binary'}
           });
         });
-        it('sets this.showing correctly', () => {
+        it('sets this.showing correctly', function() {
           delete showing.field5;
           expect(store.showing).to.deep.equal(showing);
         });
       });
     });
-    describe('isArray=true', () => {
-      describe('from middle of array', () => {
-        describe('last column is now empty', () => {
-          before((done) => {
+    describe('isArray=true', function() {
+      describe('from middle of array', function() {
+        describe('last column is now empty', function() {
+          before(function(done) {
             store = configureStore({ actions: actions });
             store.resetColumns(arrayCols);
             expect(store.columns).to.deep.equal(arrayCols);
             expect(store.showing).to.deep.equal(arrayShowing);
             done();
           });
-          after((done) => {
+          after(function(done) {
             done();
           });
-          it('does trigger with remove', (done) => {
+          it('does trigger with remove', function(done) {
             const unsubscribe = store.listen((params) => {
               unsubscribe();
               expect(params).to.deep.equal({
@@ -524,30 +524,30 @@ describe('store', () => {
             });
             store.elementRemoved(1, 'id1', true);
           });
-          it('updates this.columns correctly', () => {
+          it('updates this.columns correctly', function() {
             expect(store.columns).to.deep.equal({
               0: {id1: 'String', id2: 'Int64'},
               1: {id1: 'Int32', id2: 'Double'}
             });
           });
-          it('updates this.showing correctly', () => {
+          it('updates this.showing correctly', function() {
             expect(store.showing).to.deep.equal({
               0: 'Mixed', 1: 'Mixed'
             });
           });
         });
-        describe('last column is not empty', () => {
-          before((done) => {
+        describe('last column is not empty', function() {
+          before(function(done) {
             store = configureStore({ actions: actions });
             store.resetColumns(arrayCols);
             expect(store.columns).to.deep.equal(arrayCols);
             expect(store.showing).to.deep.equal(arrayShowing);
             done();
           });
-          after((done) => {
+          after(function(done) {
             done();
           });
-          it('does not trigger with remove', (done) => {
+          it('does not trigger with remove', function(done) {
             const unsubscribe = store.listen((params) => {
               unsubscribe();
               expect(params).to.deep.equal({
@@ -558,33 +558,33 @@ describe('store', () => {
             });
             store.elementRemoved(0, 'id2', true);
           });
-          it('updates this.columns correctly', () => {
+          it('updates this.columns correctly', function() {
             expect(store.columns).to.deep.equal({
               0: {id1: 'String', id2: 'Double'},
               1: {id1: 'Int64'},
               2: {id1: 'Int32'}
             });
           });
-          it('updates this.showing correctly', () => {
+          it('updates this.showing correctly', function() {
             expect(store.showing).to.deep.equal({
               0: 'Mixed', 1: 'Int64', 2: 'Int32'
             });
           });
         });
       });
-      describe('from end of array', () => {
-        describe('last column is now empty', () => {
-          before((done) => {
+      describe('from end of array', function() {
+        describe('last column is now empty', function() {
+          before(function(done) {
             store = configureStore({ actions: actions });
             store.resetColumns(arrayCols);
             expect(store.columns).to.deep.equal(arrayCols);
             expect(store.showing).to.deep.equal(arrayShowing);
             done();
           });
-          after((done) => {
+          after(function(done) {
             done();
           });
-          it('does trigger with remove', (done) => {
+          it('does trigger with remove', function(done) {
             const unsubscribe = store.listen((params) => {
               unsubscribe();
               expect(params).to.deep.equal({
@@ -595,30 +595,30 @@ describe('store', () => {
             });
             store.elementRemoved(2, 'id1', true);
           });
-          it('updates this.columns correctly', () => {
+          it('updates this.columns correctly', function() {
             expect(store.columns).to.deep.equal({
               0: {id1: 'String', id2: 'Int64'},
               1: {id1: 'Int64', id2: 'Double'}
             });
           });
-          it('updates this.showing correctly', () => {
+          it('updates this.showing correctly', function() {
             expect(store.showing).to.deep.equal({
               0: 'Mixed', 1: 'Mixed'
             });
           });
         });
-        describe('last column is not empty', () => {
-          before((done) => {
+        describe('last column is not empty', function() {
+          before(function(done) {
             store = configureStore({ actions: actions });
             store.resetColumns(arrayCols);
             expect(store.columns).to.deep.equal(arrayCols);
             expect(store.showing).to.deep.equal(arrayShowing);
             done();
           });
-          after((done) => {
+          after(function(done) {
             done();
           });
-          it('does not trigger with remove', (done) => {
+          it('does not trigger with remove', function(done) {
             const unsubscribe = store.listen((params) => {
               unsubscribe();
               expect(params).to.deep.equal({
@@ -629,14 +629,14 @@ describe('store', () => {
             });
             store.elementRemoved(1, 'id2', true);
           });
-          it('updates this.columns correctly', () => {
+          it('updates this.columns correctly', function() {
             expect(store.columns).to.deep.equal({
               0: {id1: 'String', id2: 'Int64'},
               1: {id1: 'Int64'},
               2: {id1: 'Int32'}
             });
           });
-          it('updates this.showing correctly', () => {
+          it('updates this.showing correctly', function() {
             expect(store.showing).to.deep.equal({
               0: 'Mixed', 1: 'Int64', 2: 'Int32'
             });
@@ -644,9 +644,9 @@ describe('store', () => {
         });
       });
     });
-    describe('with staged elements', () => {
-      describe('element being removed was staged', () => {
-        before((done) => {
+    describe('with staged elements', function() {
+      describe('element being removed was staged', function() {
+        before(function(done) {
           store = configureStore({ actions: actions });
           store.resetColumns(arrayCols);
           store.elementMarkRemoved(1, 'id1');
@@ -663,10 +663,10 @@ describe('store', () => {
           });
           done();
         });
-        after((done) => {
+        after(function(done) {
           done();
         });
-        it('does trigger with remove', (done) => {
+        it('does trigger with remove', function(done) {
           const unsubscribe = store.listen((params) => {
             unsubscribe();
             expect(params).to.deep.equal({
@@ -680,24 +680,24 @@ describe('store', () => {
           });
           store.elementRemoved(1, 'id1', true);
         });
-        it('updates this.columns correctly', () => {
+        it('updates this.columns correctly', function() {
           expect(store.columns).to.deep.equal({
             0: {id1: 'String', id2: 'Int64'},
             1: {id1: 'Int32', id2: 'Double'}
           });
         });
-        it('updates this.showing correctly', () => {
+        it('updates this.showing correctly', function() {
           expect(store.showing).to.deep.equal({
             0: 'Mixed', 1: 'Mixed'
           });
         });
-        it('updates this.stageRemoved correctly', () => {
+        it('updates this.stageRemoved correctly', function() {
           expect(store.stageRemove).to.deep.equal({});
         });
       });
-      describe('element being removed was not staged', () => {
-        describe('element marked as removed is after element removed', () => {
-          before((done) => {
+      describe('element being removed was not staged', function() {
+        describe('element marked as removed is after element removed', function() {
+          before(function(done) {
             store = configureStore({ actions: actions });
             store.resetColumns(arrayCols);
             store.elementMarkRemoved(2, 'id1');
@@ -713,10 +713,10 @@ describe('store', () => {
             });
             done();
           });
-          after((done) => {
+          after(function(done) {
             done();
           });
-          it('does trigger with remove', (done) => {
+          it('does trigger with remove', function(done) {
             const unsubscribe = store.listen((params) => {
               unsubscribe();
               expect(params).to.deep.equal({
@@ -730,18 +730,18 @@ describe('store', () => {
             });
             store.elementRemoved(1, 'id1', true);
           });
-          it('updates this.columns correctly', () => {
+          it('updates this.columns correctly', function() {
             expect(store.columns).to.deep.equal({
               0: {id1: 'String', id2: 'Int64'},
               1: {id2: 'Double'}
             });
           });
-          it('updates this.showing correctly', () => {
+          it('updates this.showing correctly', function() {
             expect(store.showing).to.deep.equal({
               0: 'Mixed', 1: 'Double'
             });
           });
-          it('updates this.stageRemoved correctly', () => {
+          it('updates this.stageRemoved correctly', function() {
             expect(store.stageRemove).to.deep.equal({
               1: {id1: true}
             });
@@ -751,14 +751,14 @@ describe('store', () => {
     });
   });
 
-  describe('#elementTypeChanged', () => {
-    before((done) => {
+  describe('#elementTypeChanged', function() {
+    before(function(done) {
       store = configureStore({ actions: actions });
       store.resetColumns({field1: {3: 'Object'}});
       done();
     });
-    describe('Changing the type of the last item', () => {
-      it('triggers correctly', (done) => {
+    describe('Changing the type of the last item', function() {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({updateHeaders: {showing: {field1: 'Date'}}});
@@ -766,35 +766,35 @@ describe('store', () => {
         });
         store.elementTypeChanged('field1', 'Date', '3');
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({field1: {3: 'Date'}});
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal({field1: 'Date'});
       });
     });
-    describe('Casted to the same type', () => {
-      it('does not trigger', () => {
+    describe('Casted to the same type', function() {
+      it('does not trigger', function() {
         const unsubscribe = store.listen(() => {
           assert.fail();
         });
         store.elementTypeChanged('field1', 'Date', '3');
         unsubscribe();
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({field1: {3: 'Date'}});
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal({field1: 'Date'});
       });
     });
-    describe('Changing the type to Mixed', () => {
-      before((done) => {
+    describe('Changing the type to Mixed', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         store.resetColumns({field1: {3: 'Date', 2: 'Date'}});
         done();
       });
-      it('triggers correctly', (done) => {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({updateHeaders: {showing: {field1: 'Mixed'}}});
@@ -802,15 +802,15 @@ describe('store', () => {
         });
         store.elementTypeChanged('field1', 'Double', '3');
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({field1: {3: 'Double', 2: 'Date'}});
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal({field1: 'Mixed'});
       });
     });
-    describe('Changing the type from Mixed', () => {
-      it('triggers correctly', (done) => {
+    describe('Changing the type from Mixed', function() {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({updateHeaders: {showing: {field1: 'Double'}}});
@@ -818,17 +818,17 @@ describe('store', () => {
         });
         store.elementTypeChanged('field1', 'Double', '2');
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({field1: {3: 'Double', 2: 'Double'}});
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal({field1: 'Double'});
       });
     });
   });
 
-  describe('#elementMarkRemoved', () => {
-    before((done) => {
+  describe('#elementMarkRemoved', function() {
+    before(function(done) {
       store = configureStore({ actions: actions });
       store.resetColumns(columns);
       showing.field1 = 'Object';
@@ -838,8 +838,8 @@ describe('store', () => {
       expect(store.stageRemove).to.deep.equal({});
       done();
     });
-    describe('marking an element as removed with the same type', () => {
-      it('does not trigger', () => {
+    describe('marking an element as removed with the same type', function() {
+      it('does not trigger', function() {
         const unsubscribe = store.listen(() => {
           assert.fail();
         });
@@ -848,7 +848,7 @@ describe('store', () => {
         store.elementMarkRemoved('field1', '3');
         unsubscribe();
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
           field3: {1: 'Int32', 2: 'Int32', 3: 'Int32'},
@@ -856,22 +856,22 @@ describe('store', () => {
           field5: {1: 'Int32', 2: 'Int32', 3: 'String'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets this.stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field1: {1: true, 2: true, 3: true}});
       });
     });
-    describe('marking an element as removed with a different type but still Mixed', () => {
-      it('does not trigger', () => {
+    describe('marking an element as removed with a different type but still Mixed', function() {
+      it('does not trigger', function() {
         const unsubscribe = store.listen(() => {
           assert.fail();
         });
         store.elementMarkRemoved('field5', '1');
         unsubscribe();
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
           field3: {1: 'Int32', 2: 'Int32', 3: 'Int32'},
@@ -879,15 +879,15 @@ describe('store', () => {
           field5: {2: 'Int32', 3: 'String'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets this.stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field1: {1: true, 2: true, 3: true}, field5: {1: true}});
       });
     });
-    describe('marking an element as removed with a different type no longer Mixed', () => {
-      it('triggers correctly', (done) => {
+    describe('marking an element as removed with a different type no longer Mixed', function() {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({updateHeaders: {showing: {field5: 'Int32'}}});
@@ -895,7 +895,7 @@ describe('store', () => {
         });
         store.elementMarkRemoved('field5', '3');
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
           field3: {1: 'Int32', 2: 'Int32', 3: 'Int32'},
@@ -903,38 +903,38 @@ describe('store', () => {
           field5: {2: 'Int32'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         showing.field5 = 'Int32';
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets this.stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field1: {1: true, 2: true, 3: true}, field5: {1: true, 3: true}});
       });
     });
-    describe('marking the last element in a column as removed', () => {
-      it('does not trigger', () => {
+    describe('marking the last element in a column as removed', function() {
+      it('does not trigger', function() {
         const unsubscribe = store.listen(() => {
           assert.fail();
         });
         store.elementMarkRemoved('field5', '2');
         unsubscribe();
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
           field3: {1: 'Int32', 2: 'Int32', 3: 'Int32'},
           field4: {1: 'Binary', 2: 'Binary', 3: 'Binary'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets this.stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field1: {1: true, 2: true, 3: true}, field5: {1: true, 2: true, 3: true}});
       });
     });
-    describe('calling elementAdded for a marked removed element adds it back', () => {
-      it('triggers correctly', (done) => {
+    describe('calling elementAdded for a marked removed element adds it back', function() {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({updateHeaders: {showing: {field1: 'String'}}});
@@ -942,7 +942,7 @@ describe('store', () => {
         });
         store.elementAdded('field1', 'String', '1');
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field1: {1: 'String'},
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
@@ -950,16 +950,16 @@ describe('store', () => {
           field4: {1: 'Binary', 2: 'Binary', 3: 'Binary'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         showing.field1 = 'String';
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets this.stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field1: {2: true, 3: true}, field5: {1: true, 2: true, 3: true}});
       });
     });
-    describe('calling elementRemoved for a marked remove element removes it', () => {
-      it('does not trigger if there are marked removed elements in the column', () => {
+    describe('calling elementRemoved for a marked remove element removes it', function() {
+      it('does not trigger if there are marked removed elements in the column', function() {
         const unsubscribe = store.listen(() => {
           assert.fail();
         });
@@ -967,21 +967,21 @@ describe('store', () => {
         store.elementRemoved('field1', '2');
         unsubscribe();
       });
-      it('sets this.columns correctly', () => {
+      it('sets this.columns correctly', function() {
         expect(store.columns).to.deep.equal({
           field2: {1: 'Array', 2: 'Array', 3: 'Array'},
           field3: {1: 'Int32', 2: 'Int32', 3: 'Int32'},
           field4: {1: 'Binary', 2: 'Binary', 3: 'Binary'}
         });
       });
-      it('sets this.showing correctly', () => {
+      it('sets this.showing correctly', function() {
         showing.field5 = 'Int32';
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets this.stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field1: {3: true}, field5: {1: true, 2: true, 3: true}});
       });
-      it('triggers if there are no  more marked removed elements in the column', (done) => {
+      it('triggers if there are no  more marked removed elements in the column', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({remove: {colIds: ['field1']}});
@@ -989,26 +989,19 @@ describe('store', () => {
         });
         store.elementRemoved('field1', '3');
       });
-      it('sets this.columns correctly', () => {
-        expect(store.columns).to.deep.equal({
-          field2: {1: 'Array', 2: 'Array', 3: 'Array'},
-          field3: {1: 'Int32', 2: 'Int32', 3: 'Int32'},
-          field4: {1: 'Binary', 2: 'Binary', 3: 'Binary'}
-        });
-      });
-      it('sets this.showing correctly', () => {
+      it('sets field1 this.showing correctly', function() {
         delete showing.field1;
         expect(store.showing).to.deep.equal(showing);
       });
-      it('sets this.stageRemove correctly', () => {
+      it('sets stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({field5: {1: true, 2: true, 3: true}});
       });
     });
   });
 
-  describe('#cleanCols', () => {
-    describe('columns have elements marked for deletion', () => {
-      before((done) => {
+  describe('#cleanCols', function() {
+    describe('columns have elements marked for deletion', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         store.resetColumns({field1: {1: 'Object', 2: 'Object', 3: 'Object'}});
         store.elementMarkRemoved('field1', 1);
@@ -1019,21 +1012,21 @@ describe('store', () => {
         expect(store.stageRemove).to.deep.equal({field1: {1: true, 2: true, 3: true}});
         done();
       });
-      it('does not trigger', () => {
+      it('does not trigger', function() {
         const unsubscribe = store.listen(() => {
           assert.fail();
         });
         store.cleanCols();
         unsubscribe();
       });
-      it('does not make internal changes', () => {
+      it('does not make internal changes', function() {
         expect(store.columns).to.deep.equal({});
         expect(store.showing).to.deep.equal({field1: 'Object'});
         expect(store.stageRemove).to.deep.equal({field1: {1: true, 2: true, 3: true}});
       });
     });
-    describe('all elements in column are removed', () => {
-      before((done) => {
+    describe('all elements in column are removed', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         store.resetColumns({field1: {3: 'Object'}, field2: {3: 'Array'}});
         delete store.columns.field1;
@@ -1043,7 +1036,7 @@ describe('store', () => {
         expect(store.stageRemove).to.deep.equal({});
         done();
       });
-      it('triggers correctly', (done) => {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({remove: {colIds: ['field1', 'field2']}});
@@ -1051,7 +1044,7 @@ describe('store', () => {
         });
         store.cleanCols();
       });
-      it('makes internal changes', () => {
+      it('makes internal changes', function() {
         expect(store.columns).to.deep.equal({});
         expect(store.showing).to.deep.equal({});
         expect(store.stageRemove).to.deep.equal({});
@@ -1059,17 +1052,17 @@ describe('store', () => {
     });
   });
 
-  describe('#renameColumn', () => {
-    before((done) => {
+  describe('#renameColumn', function() {
+    before(function(done) {
       store = configureStore({ actions: actions });
       store.resetColumns({$new: {id1: 'Int32', id2: 'String'}});
       store.elementMarkRemoved('$new', 'id2');
       done();
     });
-    after((done) => {
+    after(function(done) {
       done();
     });
-    it('updates objects', () => {
+    it('updates objects', function() {
       store.renameColumn('$new', 'fieldname');
       expect(store.columns).to.deep.equal({fieldname: {id1: 'Int32'}});
       expect(store.showing).to.deep.equal({fieldname: 'Int32'});
@@ -1077,9 +1070,9 @@ describe('store', () => {
     });
   });
 
-  describe('#replaceDoc', () => {
-    describe('replacing a document with a smaller document', () => {
-      before((done) => {
+  describe('#replaceDoc', function() {
+    describe('replacing a document with a smaller document', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         store.resetColumns({
           0: {id1: 'String', id2: 'Int32'},
@@ -1089,10 +1082,10 @@ describe('store', () => {
         store.elementMarkRemoved(0, 'id1');
         done();
       });
-      after((done) => {
+      after(function(done) {
         done();
       });
-      it('triggers correctly', (done) => {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({
@@ -1107,25 +1100,25 @@ describe('store', () => {
           'id1', 'id1', {0: 1, 1: 'a string'}
         );
       });
-      it('updates columns correctly', () => {
+      it('updates columns correctly', function() {
         expect(store.columns).to.deep.equal({
           0: {id1: 'Int32', id2: 'Int32'},
           1: {id1: 'String', id2: 'Double'}
         });
       });
-      it('updates showing correctly', () => {
+      it('updates showing correctly', function() {
         expect(store.showing).to.deep.equal({
           0: 'Int32',
           1: 'Mixed',
           2: 'Int32'
         });
       });
-      it('updates stageRemove correctly', () => {
+      it('updates stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({});
       });
     });
-    describe('replacing a document with a larger document', () => {
-      before((done) => {
+    describe('replacing a document with a larger document', function() {
+      before(function(done) {
         store = configureStore({ actions: actions });
         store.resetColumns({
           0: {id1: 'String', id2: 'Int32'},
@@ -1135,10 +1128,10 @@ describe('store', () => {
         store.elementMarkRemoved(1, 'id2');
         done();
       });
-      after((done) => {
+      after(function(done) {
         done();
       });
-      it('triggers correctly', (done) => {
+      it('triggers correctly', function(done) {
         const unsubscribe = store.listen((params) => {
           unsubscribe();
           expect(params).to.deep.equal({
@@ -1155,7 +1148,7 @@ describe('store', () => {
           }
         );
       });
-      it('updates columns correctly', () => {
+      it('updates columns correctly', function() {
         expect(store.columns).to.deep.equal({
           0: {id1: 'Int32', id2: 'Int32'},
           1: {id1: 'String'},
@@ -1164,7 +1157,7 @@ describe('store', () => {
           4: {id1: 'String'}
         });
       });
-      it('updates showing correctly', () => {
+      it('updates showing correctly', function() {
         expect(store.showing).to.deep.equal({
           0: 'Int32',
           1: 'String',
@@ -1173,7 +1166,7 @@ describe('store', () => {
           4: 'String'
         });
       });
-      it('updates stageRemove correctly', () => {
+      it('updates stageRemove correctly', function() {
         expect(store.stageRemove).to.deep.equal({1: {id2: true}});
       });
     });
