+++ packages/compass-crud/test/setup.js
@@ -1,27 +1,27 @@
-const Enzyme = require('enzyme');
-const Adapter = require('enzyme-adapter-react-16');
-Enzyme.configure({ adapter: new Adapter() });
+// const Enzyme = require('enzyme');
+// const Adapter = require('enzyme-adapter-react-16');
+// Enzyme.configure({ adapter: new Adapter() });
 
-const chai = require('chai');
-const chaiEnzyme = require('chai-enzyme');
-const sinonChai = require('sinon-chai');
+// const chai = require('chai');
+// const chaiEnzyme = require('chai-enzyme');
+// const sinonChai = require('sinon-chai');
 
-global.AbortController = window.AbortController;
-global.sinon = require('sinon');
-global.expect = chai.expect;
-global.jQuery = require('jquery');
+// global.AbortController = window.AbortController;
+// global.sinon = require('sinon');
+// global.expect = chai.expect;
+// global.jQuery = require('jquery');
 
-class Worker {
-  postMessage() {
-    return true;
-  }
-  terminate() {
-    return true;
-  }
-}
+// class Worker {
+//   postMessage() {
+//     return true;
+//   }
+//   terminate() {
+//     return true;
+//   }
+// }
 
-global.Worker = Worker;
+// global.Worker = Worker;
 
-chai.should();
-chai.use(sinonChai);
-chai.use(chaiEnzyme());
+// chai.should();
+// chai.use(sinonChai);
+// chai.use(chaiEnzyme());
